#451: Move PageError to common
